Update intro_getting_started.rst
New to ansible, but installing via HomeBrew on Mac OS X (Mountain Lion), I discovered from a note on http://lampros.chaidas.com/index.php/2014/06/22/beginning-with-ansible-on-freebsd-10/ that the inventory file (hosts) needs to be in /usr/local/etc/ansible not /etc/ansible. Will submit change for Inventory doc as well. With a hosts file in /etc/ansible, ansible did not detect it, and continued to prompt me to specify one with the "-i" parameter. Am I wrong, or is this an OS X anomaly?